pochemy to ne rabotaet?

+++ Ant.java
@@ -4,15 +4,15 @@ import java.awt.image.BufferedImage;
 import javax.swing.JFrame;
 
 abstract public class Ant {
-    protected int x;
-    protected int y;
+    protected double x;
+    protected double y;
     int birthTime;
     int lifeTime;
     int id;
 
     boolean returning; // Добавляем поле для отслеживания направления движения муравья
     double angle; // Для муравьев-воинов, чтобы хранить текущий угол движения
-    int speed; // Для муравьев-воинов, чтобы хранить скорость движения
+    double speed; // Для муравьев-воинов, чтобы хранить скорость движения
 
 
     //protected BufferedImage icon;
